requirements.txt
torch==1.9.0+cpu
torchvision==0.10.0+cpu
-f https://download.pytorch.org/whl/torch_stable.html
transformers==4.11.3
ultralytics==6.2
pillow==8.3.2
fastapi==0.68.1
uvicorn==0.15.0